chore: move classes into their own files
* Adds an eslint rule to check for this behaviour
* Switch to interface for `AbstractDBClient` which better reflects
its usage
* Refactor AbstractDatabase to align with Abstract class conventions
