+++ .eslintrc.json
@@ -46,6 +46,7 @@
     "react/prop-types": "off",
     "react/no-unknown-property": "warn",
     "@typescript-eslint/no-var-requires": "off",
-    "@typescript-eslint/no-namespace": "off"
+    "@typescript-eslint/no-namespace": "off",
+    "max-classes-per-file": "error"
   }
 }
