+++ src/stores/databaseV2/clients/firestore.tsx
@@ -1,6 +1,7 @@
 import { IDBEndpoint, DBDoc } from 'src/models/common.models'
 import { firestore } from 'src/utils/firebase'
-import { DBQueryOptions, AbstractDBClient } from '../types'
+import { DBQueryOptions } from '../types'
+import { AbstractDBClient } from "../types"
 import { Observable, Observer } from 'rxjs'
 import { DB_QUERY_DEFAULTS } from '../utils/db.utils'
