+++ src/stores/databaseV2/types/index.d.ts
@@ -4,40 +4,7 @@
  * API items.  On a detail page for a single item, the summary will be
  * shown followed by the remarks section (if any).
  */
-import { Observable } from 'rxjs'
-export abstract class AbstractDatabase {
-  /**
-   * Gets a `CollectionReference` instance that refers to the collection at
-   * the specified path.
-   *
-   * @param clients pass a `DBClients` object to specify what to use as
-   * server and cache dbs. Uses a default configuration if not specified
-   */
-  constructor(clients?: DBClients)
-  /**
-   * Gets a `CollectionReference` instance that refers to the collection at
-   * the specified path.
-   *
-   * @param endpoint The collection/table path name
-   * @return The `CollectionReference` instance.
-   */
-  collection(endpoint: DBEndpoint): CollectionReference
-}
-
-export class CollectionReference {
-  /**
-   * The `clients` is passed to collection references so that the collection
-   * has full access to required db client methods (e.g. server and cache)
-   */
-  private constructor(endpoint: DBEndpoint, clients: DBClients)
-
-  /**
-   *
-   * @param docID pass a `docID` to provide a reference to a specific database document
-   * Leaving blank will generate a new id
-   */
-  doc(docID?: string): DocReference
-}
+export type { AbstractDBClient } from './DBClient'
 
 /**
  * The `DBClients` consists of separate databases for use online and offline.
@@ -53,30 +20,6 @@ export interface DBClients {
   cacheDB: AbstractDBClient
   serverDB: AbstractDBClient
   serverCacheDB: AbstractDBClient
-}
-
-export abstract class AbstractDBClient {
-  getDoc<T>(endpoint: string, docId: string): Promise<(T & DBDoc) | undefined>
-
-  setDoc(endpoint: string, doc: any): Promise<void>
-
-  setBulkDocs(endpoint: string, docs: any): Promise<void>
-
-  getCollection<T>(endpoint: string): Promise<(T & DBDoc)[]>
-
-  queryCollection<T>(
-    endpoint: string,
-    queryOpts: DBQueryOptions,
-  ): Promise<(T & DBDoc)[]>
-
-  streamCollection?<T>(
-    endpoint: string,
-    queryOpts?: DBQueryOptions,
-  ): Observable<(T & DBDoc)[]>
-
-  streamDoc?<T>(endpoint: string): Observable<T & DBDoc>
-
-  deleteDoc(endpoint: string, docId: string): Promise<void>
 }
 
 /**
